Made Changes to checkMove
Code for CheckMove has been written, and should work, however I'm
running into an error with the first line as it is not getting past it
at all, which is whay any calls to it have been commented out.

If you remove that first line it runs as it should, however it does not
catch errors where the space chosen to move to or from is not a possible
space to move to.