Actually fixed the animations and have a game that (mostly) doesnt slow down now
For whatever reason, the fire animations are the only ones that slow down the game.

Not sure why as they run the same as the ice animations in 3DS Max...

But it will cycle through the animations now, sometimes it plays the same one twice but it will play all of them eventually it seems.